# docs/openapi/pn-radd-alt-internal.yaml
openapi: 3.0.1
info:
  termsOfService: https://notifichedigitali.pagopa.it/radd/index.html
  x-api-id: api-internal-radd-net # NO EXTERNAL
  #  x-api-id: api-external-radd-net # ONLY EXTERNAL
  title: 'Piattaforma Notifiche: API B2B per gli operatori RADD'
  x-summary: 'Piattaforma Notifiche: API B2B per gli operatori RADD'
  version: '1.0.0'
  description: >
    # Abstract
    
      API utilizzate dai soggetti abilitati (Operatori RADD) all’utilizzo dei
      sistemi **RADD** (Rete di Assorbimento del Divario Digitale) per rendere
      fruibile **SEND** anche a destinatari in condizione di Digital Divide,
      intesi come non dotati di **SPID** o di **CIE**, che per questo non possono
      accedere alla notifica tramite portale web o per mezzo di **app IO**. 
    
    
      Esistono due casi d’uso **RADD**:  
    
    
      - **ACT**: Richiesta degli atti di una determinata notifica tramite 
    i dati presenti nell’avviso di avvenuta ricezione (AAR). 
    
    
      - **AOR**: Richiesta degli avvisi di avvenuta ricezione (AAR) non
      recapitati per irreperibilità.
    
    
      # Autenticazione 
    
      L'ente che si integra con le API REST della RADD deve autenticarsi tramite
      JSON Web Token (JWT) firmato con algoritmo a chiave asimmetrica RS256. 
    
      Il token deve essere passato sul header Authorization della richiesta nella
      forma: `Authorization: Bearer < token >`
    
      ## Identificativo utente
    
...
    
      # CASO D’USO ACT
    
...

...
servers:
...
paths:
  /radd-net/api/v1/documents/upload:
    ## SLA - Tempo risposta medio: < 1s; Timeout :3s
    post:
      tags:
        - documentOperations
      description: >-
        API utilizzata per la richiesta della presigned URL utilizzata per il caricamento della scansione
        dei documenti e metadati presentati a sportello dal cliente. Il file deve essere in formato ZIP
        e la struttura dello zip deve seguire quanto specificato in documentazione.
      operationId: documentUpload
      parameters:
        - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
        - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
        - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
        - $ref: '#/components/parameters/uid'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/download/{operationType}/{operationId}:
    ## SLA - Tempo risposta medio: < 10s; Timeout :20s
    parameters:
      - $ref: '#/components/parameters/operationType'
      - $ref: '#/components/parameters/operationId'
      - $ref: '#/components/parameters/attachmentId'
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
    get:
      tags:
        - documentOperations
      operationId: documentDownload
      description: API utilizzata per il download del frontespizio
      responses:
        200:
          description: Ritorna il PDF del frontespizio
          content:
            'application/pdf':
              schema:
#               type: object    # ONLY EXTERNAL
                type: string    # NO EXTERNAL
                format: byte    # NO EXTERNAL
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/act/inquiry:
    ## SLA - Tempo risposta medio: < 1s; Timeout :3s
    parameters:
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
      - $ref: '#/components/parameters/uid'
      - $ref: '#/components/parameters/recipientTaxId'
      - $ref: '#/components/parameters/recipientType'
      - $ref: '#/components/parameters/qrCode'
      - $ref: '#/components/parameters/iun'
    get:
      tags:
        - actOperations
      operationId: actInquiry
      x-pagopa-pn-custom-throttling-group: 'radd-by-cf'
      x-pagopa-pn-custom-throttling-value: 'method.request.querystring.recipientTaxId'
      description: API utilizzata per la verifica della presenza a sistema di atti e attestazioni tramite iun o qrCode.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActInquiryResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/aor/inquiry:
    ## SLA - Tempo risposta medio: < 1s; Timeout :3s"
    parameters:
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
      - $ref: '#/components/parameters/uid'
      - $ref: '#/components/parameters/recipientTaxId'
      - $ref: '#/components/parameters/recipientType'
    get:
      tags:
        - aorOperations
      operationId: aorInquiry
      x-pagopa-pn-custom-throttling-group: 'radd-by-cf'
      x-pagopa-pn-custom-throttling-value: 'method.request.querystring.recipientTaxId'
      description: API utilizzata per la verifica della presenza a sistema di avvisi di avvenuta ricezione per il cliente.
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AORInquiryResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/act/transaction/start:
    ## SLA - Tempo risposta medio: < 10s; Timeout: 20s. Tempo massimo polling (response.status.code = 0): 10s
    parameters:
      - $ref: '#/components/parameters/uid'
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
    post:
      tags:
        - actOperations
      operationId: startActTransaction
      x-pagopa-pn-custom-throttling-group: 'radd-by-cf'
      x-pagopa-pn-custom-throttling-value: 'method.request.body.recipientTaxId'
      description: API utilizzata per la richiesta di avvio dell'operazione di scaricamento e consegna dell'atto e delle eventuali attestazioni.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActStartTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/act/transaction/complete:
    ## SLA - Tempo risposta medio: < 1s, Timeout :3s
    parameters:
      - $ref: '#/components/parameters/uid'
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
    post:
      tags:
        - actOperations
      operationId: completeActTransaction
      description: API utilizzata per la notifica del completamento dell'operazione di consegna dell'atto e delle eventuali attestazioni.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/act/transaction/abort:
    ## SLA - Tempo risposta medio: < 1s; Timeout :3s
    parameters:
      - $ref: '#/components/parameters/uid'
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
    post:
      tags:
        - actOperations
      operationId: abortActTransaction
      description: API utilizzata per la notifica  dell'annullamento dell'operazione di consegna dell'atto e delle eventuali attestazioni.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbortTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/aor/transaction/start:
    ## SLA - Tempo risposta medio: < 1s; Timeout: 3s. Tempo massimo polling(response.status.code = 0): 10s
    parameters:
      - $ref: '#/components/parameters/uid'
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: '#/components/parameters/cxRoleAuthFleet'                       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
    post:
      tags:
        - aorOperations
      operationId: startAorTransaction
      x-pagopa-pn-custom-throttling-group: 'radd-by-cf'
      x-pagopa-pn-custom-throttling-value: 'method.request.body.recipientTaxId'
      description: API utilizzata per la richiesta di avvio dell'operazione di scaricamento e consegna degli avvisi di avvenuta ricezione.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AorStartTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/aor/transaction/complete:
    ## SLA - Tempo risposta medio: < 1s; Timeout :3s
    parameters:
      - $ref: '#/components/parameters/uid'
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
    post:
      tags:
        - aorOperations
      operationId: completeAorTransaction
      description: API utilizzata per la notifica  del completamento dell'operazione di consegna dell'atto.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompleteTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
  /radd-net/api/v1/aor/transaction/abort:
    ## SLA - Tempo risposta medio: < 1s; Timeout :3s
    parameters:
      - $ref: '#/components/parameters/uid'
      - $ref: 'remote-refs.yaml#/components/parameters/cxTypeAuthFleet'       # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/cxIdAuthFleet'         # NO EXTERNAL
      - $ref: 'remote-refs.yaml#/components/parameters/headerSourceChannel'   # NO EXTERNAL
    post:
      tags:
        - aorOperations
      operationId: abortAorTransaction
      description: API utilizzata per la notifica dell'annullamento dell'operazione di consegna del/degli avviso/i di avvenuta ricezione.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbortTransactionRequest'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbortTransactionResponse'
        400:
          $ref: '#/components/responses/Invalid'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        405:
          $ref: '#/components/responses/NotAllowed'
        500:
          $ref: '#/components/responses/InternalError'
components:
  schemas:
    DocumentUploadRequest:
      description: Richiesta di upload di un file zip contenente i documenti/metadati verso PagoPA
      type: object
      required:
        - operationId
        - checksum
      properties:
        operationId:
          type: string
        checksum:
          type: string
          example: "708F4C8216F30FA6007F8E2F316ECC935D94057202FC5D8008BCCC118EA12560"
    DocumentUploadResponse:
      description: Response della Richiesta di upload
      type: object
      properties:
        url:
          type: string
          example: "https://presignedurldemo.s3.eu-west-2.amazonaws.com/image.png?X-Amz-Algorithm=AWS4-HMAC-SHA256&X-Amz-Credential=AKIAJJWZ7B6WCRGMKFGQ%2F20180210%2Feu-west-2%2Fs3%2Faws4_request&X-Amz-Date=20180210T171315Z&X-Amz-Expires=1800&X-Amz-Signature=12b74b0788aa036bc7c3d03b3f20c61f1f91cc9ad8873e3314255dc479a25351&X-Amz-SignedHeaders=host"
        fileKey:
          type: string
          example: "PN_RADD_ALT_ATTACHMENT-4958442ee6eb467abec049afa67e3899.zip"
        secret:
          type: string
          example: "alow22ddw"
        status:
          $ref: '#/components/schemas/ResponseStatus'
    ActInquiryResponse:
      description: Response di verifica presenza atto ed eventuali attestazione opponibili a terzi
      type: object
      properties:
        result:
          type: boolean
        status:
          $ref: '#/components/schemas/ActInquiryResponseStatus'
    AORInquiryResponse:
      description: Response di verifica presenza AAR.
      type: object
      properties:
        result:
          type: boolean
          example: true
        status:
          $ref: '#/components/schemas/ResponseStatus'
    ActStartTransactionRequest:
      description: Richiesta  di completamento operazione.
      allOf:
        - $ref: '#/components/schemas/AorStartTransactionRequest'
      properties:
        qrCode:
          description: in alternativa allo iun
          type: string
        iun:
          description: in alternativa al qrCode
          type: string
    StartTransactionResponse:
      description: Response alla richiesta di avvio della transazione
      type: object
      properties:
        downloadUrlList:
          type: array
          items:
            $ref: '#/components/schemas/DownloadUrl'
        status:
          $ref: '#/components/schemas/StartTransactionResponseStatus'
...
    AorStartTransactionRequest:
      description: Richiesta  di completamento operazione
      type: object
      required:
        - operationId
        - recipientTaxId
        - recipientType
      properties:
        versionToken:
          type: string
          description: valore della versione del file restituito dall'operazione PUT di updload (obbligatorio solo in caso di upload del file zip).
          example: "OXSnS_28NF1sKiaZsniWHbqwX2LktjCd"
        operationId:
          type: string
          description: identificativo univoco dell'operazione all'interno del dominio dell'operatore RADD
          example: "5d2d53ad-6306-40fe-8126-51ccdd917cef"
        recipientTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "TNTGTR76E21H751S"
        recipientType:
          type: string
          enum: ["PF", "PG"]
          example: "PF"
        delegateTaxId:
          x-field-extra-annotation: "@lombok.ToString.Exclude" # NO EXTERNAL
          type: string
          example: "RFRGRZ66E21H751B"
        fileKey:
          type: string
          example: "PN_RADD_ALT_ATTACHMENT-17a4d40e08d8498a9d78f495c6bfd56a.zip"
        checksum:
          type: string
          example: "Kg2Uwyt7rYE4II1giJwh2MtaNtq6Obup2eDE0kAps+c="
          description: Checksum SHA256 del file caricato nella chiamata di documetUpload codificato in base64
        operationDate:
          type: string
          format: date-time
          description: Data/ora di inizio della transazione in formato RFC3339.
          example: "2022-06-21T11:44:28Z"
    CompleteTransactionRequest:
      description: Richiesta  di chiusura della transazione
      type: object
      required:
        - operationId
        - operationDate
      properties:
        operationId:
          type: string
          description: identificativo univoco dell'operazione all'interno del dominio dell'operatore RADD
          example: "5d2d53ad-6306-40fe-8126-51ccdd917cef"
        operationDate:
          type: string
          format: date-time
          description: Data/ora di completamento della transazione,  espressa in formato RFC3339.
          example: "2022-06-21T11:44:28Z"
    CompleteTransactionResponse:
      description: Response alla richiesta di chiusura della transazione
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TransactionResponseStatus'
    AbortTransactionRequest:
      description: Richiesta  di cancellazione della transazione
      type: object
      required:
        - operationId
        - operationDate
      properties:
        operationId:
          type: string
          description: identificativo univoco dell'operazione all'interno del dominio dell'operatore RADD
          example: "5d2d53ad-6306-40fe-8126-51ccdd917cef"
        reason:
          type: string
          default: "abort"
        operationDate:
          type: string
          format: date-time
          description: Data/ora di annullamento della transazione, espressa in formato RFC3339.
          example: "2022-06-21T11:44:28Z"
    AbortTransactionResponse:
      description: Response alla richiesta di chiusura della transazione
      type: object
      properties:
        status:
          $ref: '#/components/schemas/TransactionResponseStatus'
    ResponseStatus:
      description: Risposta di stato restituita al chiamante
      type: object
      properties:
        code:
          description: |
            Return code.
            * 0 -> OK
            * 99 -> KO
          type: number
          enum: [0, 99]
        message:
          type: string
          example: OK
...
    ActInquiryResponseStatus:
      description: Risposta di stato restituita al chiamante
      allOf:
        - $ref: '#/components/schemas/ResponseStatus'
        - type: object
          properties:
            code:
              description: |
                Return code.
                * 0 -> OK
                * 2 -> NotReady (documenti temporaneamente non disponibili per ragioni tecniche, ritentare dopo il tempo specificato nel retryAfter)
                * 3 -> Stampa già eseguita
                * 4 -> Documento non più disponibili
                * 10 -> Input non valido
                * 80 -> Notifica annullata dalla PA
                * 99 -> KO generico
              type: number
              enum: [0,2,3,4,10,80,99]
    ErrorResponse:
      description: Risposta di errore
      properties:
        type:
          type: string
          description: descrizione tipologia errore
        status:
          type: number
          description: codice http dell'errore
        title:
          type: string
          description: descrizione errore
        detail:
          type: string
          description: dettaglio errore
        traceId:
          type: string
          description: id traccia
        timestamp:
          type: string
          description: timestamp dell'evento
        errors:
          type: array
          description: dettagli relativi all'errore
          items:
            type: string
  parameters:
    uid:
      name: uid
      description: Identificativo pseudo-anonimizzato dell'operatore RADD deve contenere l'id utente e l'id sede che corrisponde al campo externalCode dell’anagrafica delle sedi separati dal carattere @.
      in: header
      required: true
      schema:
        type: string
      example: USR1029@UFFRO129
    recipientTaxId:
      name: recipientTaxId
      description: Codice Fiscale Destinatario
      in: query
      required: true
      schema:
        type: string
      example: TNTGTR76E21H751S
    recipientType:
      name: recipientType
      description: Sigla, Persona Fisica o Persona Giuridica
      in: query
      required: true
      schema:
        type: string
        enum: ["PF", "PG"]
      example: PF
...
    qrCode:
      name: qrCode
      description: QrCode presente sull'Avviso di Avvenuta Ricezione (AAR). Da usare in alternativa a "iun".
      in: query
      required: false
      schema:
        type: string
      example: "UkdESC1BS01QLVFWWE4tMjAyMzA4LUctMV9QRi1lZjJkZWI0NS0zYzBlLTRiY2UtYWRjMC1hODFlN2Q2YjM1NGVfY2M5ZDkyMTktMWQxMi00ZTBhLWJlZDQtNmE1YWMyYzlhZmZh"
    iun:
      name: iun
      description: Identificativo Univoco della Notifica. Da usare in alternativa a "qrCode".
      in: query
      required: false
      schema:
        type: string
      example: "RGDH-AKMP-QVXN-202308-G-1"
    operationId:
      name: operationId
      description: identificativo univoco dell'operazione all'interno del dominio dell'operatore RADD
      in: path
      required: true
      schema:
        type: string
      example: "5d2d53ad-6306-40fe-8126-51ccdd917cef"
    operationType:
      name: operationType
      description: Tipo di operazione aor o act
      in: path
      required: true
      schema:
        type: string
        enum: ["aor", "act"]
      example: aor
    attachmentId:
      name: attachmentId
      description: Identificativo file
      in: query
      required: false
      schema:
        type: string
    cxRoleAuthFleet:
      name: x-pagopa-pn-cx-role
      in: header
      description: User role
      required: false
      schema:
        type: string
        # ASCII printable characters
        pattern: ^[ -~]*$
        maxLength: 64
  responses:
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Invalid:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotAllowed:
      description: Method not allowed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

#  securitySchemes: # ONLY EXTERNAL
#    bearerAuth: # ONLY EXTERNAL
#      type: http          # ONLY EXTERNAL
#      scheme: bearer      # ONLY EXTERNAL
#      bearerFormat: JWT   # ONLY EXTERNAL
#
#security: # ONLY EXTERNAL
#  - bearerAuth: [ ]     # ONLY EXTERNAL